refactor solving

+++ test/utils_tests.jl
@@ -18,7 +18,7 @@ end
 @testitem "count_shocks" begin
     import JLD2: @load
     @load "test_data/shocks.jld2" u1 u2 u3 u4
-    dx = 2π / 512
+    dx = Float32(2π / 512)
     @test RDE.count_shocks(u1, dx) == 1
     @test RDE.count_shocks(u2, dx) == 2
     @test RDE.count_shocks(u3, dx) == 3
